ADD: SE ARREGLA PROBLEMA CON LA ELIMINACION DE ANUNCIOS EXISTENTES AL MOMENTO DE REMPLAZARLOS

+++ android/papiro/src/gt/com/papiro/sincronizador/Sincronizador.java
@@ -116,6 +116,9 @@ public class Sincronizador extends AsyncTask<Void, Void, DispositivoRs>{
 	}
 
 	private void processAnuncios(List<AnuncioRs> anuncios) throws IOException {		
+		if(anuncios == null){
+			return;
+		}
 		numeroAnuncios = anuncios.size();
 		File f = new File(baseFile);
 		FileUtil.deleteRecursive(f);
